requirements.txt
transformers
underthesea
regrex
PyYAML
torch
tqdm
matplotlib
